backend/pyproject.toml
[project]
name = "smart-travel-planner"
version = "0.1.0"
description = "AI-powered travel planning agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cachetools>=7.0.6",
    "fastapi>=0.136.1",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "pgvector>=0.4.2",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "email-validator>=2.0.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.27",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.46.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

embeddings = [
    "sentence-transformers>=5.4.1",
]
